Version 3.1.0
Version 3.1.0

+++ python/createPartVectorTilePackage.py
@@ -3,6 +3,11 @@
 __author__ = 'mu_xiaoyan'
 # Version     : 2.1.0
 # Start Time  : 2017-12-07
+
+# Update Time : 2018-09-20
+# Change Log :
+## Add support for getting known XML
+
 # Update Time : 2018-8-13
 # Change Log  :
 ##  Added new logical to calculate affected bundle extent in each level to avoid vector tile lost
@@ -12,12 +17,8 @@ __author__ = 'mu_xiaoyan'
 ##  Automatically removed intermediate data and workspace
 
 
-import arcpy
-import os
-import shutil
-import time
-import zipfile
-import sys
+import time, zipfile, arcpy
+import sys, os, shutil
 
 
 # uncompress the .zip file to folder
@@ -63,10 +64,37 @@ def analysis_original_vtpk(origin_vtpk_path):
         else:
             tile_scheme = ""
             service_type = "ONLINE"
-        aux_paras = [index_polygon, tile_scheme, service_type,temp_workspace]
-        return aux_paras
+        adv_aux_paras = [index_polygon, tile_scheme, service_type,temp_workspace]
+        return adv_aux_paras
     except:
         arcpy.AddError("Original vtpk does not exist.")
+
+# Updated at 2018-9-21
+# Automatically finding xml file and allowing user to choose origin index polygon
+def get_tile_scheme_and_index(in_map):
+    try:
+        arcpy.AddMessage("# Starting finding existing XML file ...")
+        aprx = arcpy.mp.ArcGISProject("CURRENT")
+        map = aprx.listMaps(in_map)[0]
+        # getting xml file
+        map_sr = map.defaultCamera.getExtent().spatialReference
+        map_sr_name = map_sr.name
+        map_sr_wkid = map_sr.factoryCode
+        vtpk_xml_name = "VTTS_"+str(map_sr_wkid)+"_"+map_sr_name+".xml"
+        local_path = os.path.join(os.path.expanduser("~"), "AppData\Local\ESRI\Geoprocessing")
+        vtpk_xml_path = os.path.join(local_path,vtpk_xml_name)
+        arcpy.AddMessage("# Bingo! " + vtpk_xml_path + " has been found!")
+        tile_scheme = vtpk_xml_path
+        # getting service type
+        if map_sr_wkid == 3857:
+            service_type = "ONLINE"
+        else:
+            service_type = "EXISTING"
+        # building index polygons
+        common_aux_paras = [tile_scheme, service_type]
+        return common_aux_paras
+    except:
+        arcpy.AddError("Failed to get tile scheme")
 
 # Updated at 2018-8-9
 # Added new function calculating the exact bundle extent in each level of vector tile
@@ -106,7 +134,7 @@ def calculate_affected_indexes_layer(AOI,index_polygon):
                     arcpy.SelectLayerByLocation_management(index_polygon_lyr, "WITHIN", base_affected_indexes_lyr, None,
                                                            "SUBSET_SELECTION", "NOT_INVERT")
                     arcpy.CopyFeatures_management(index_polygon_lyr, bundle_polygon_each_level)
-                    arcpy.Append_management([bundle_polygon_each_level], bundle_index_polygons, "TEST")
+                    arcpy.Append_management([bundle_polygon_each_level], bundle_index_polygons, "NO_TEST")
         arcpy.Delete_management("in_memory")
         arcpy.Delete_management(AOI_lyr)
         arcpy.Delete_management(index_polygon_lyr)
@@ -127,7 +155,6 @@ def calculate_affected_indexes_layer(AOI,index_polygon):
             arcpy.AddMessage(arcpy.GetMessages())
         return ""
 
-
 def main(argv=None):
     # Input map in the current project
     in_map = arcpy.GetParameterAsText(0)
@@ -141,28 +168,42 @@ def main(argv=None):
     origin_vtpk = arcpy.GetParameterAsText(2)
     arcpy.AddMessage("# Original vtpk : {0}.".format(origin_vtpk))
 
+    # Update 2018-9-20 add boolean parameter to check if the vtpk is advanced
+    # Check if the vtpk is advanced
+    is_adv_vtpk = arcpy.GetParameter(3)
+    arcpy.AddMessage("# Is Advanced Vtpk : {0}.".format(is_adv_vtpk))
+    # Specify the index polygon
+    index_polygon = arcpy.GetParameter(4)
+    arcpy.AddMessage("# Index Polygon path : {0}.".format(index_polygon))
+
     # Specify name and workspace for new part vtpk
-    out_part_vtpk = arcpy.GetParameterAsText(3)
+    out_part_vtpk = arcpy.GetParameterAsText(5)
     arcpy.AddMessage("# New part vtpk : {0}.".format(out_part_vtpk))
 
-    execute(in_map, AOI, origin_vtpk, out_part_vtpk)
-
-
+    execute(in_map, AOI, origin_vtpk, is_adv_vtpk, index_polygon, out_part_vtpk)
 
 # Update 2018-8-10, Calling the new function
 # calculating the exact bundle extent in each level of vector tile
-def execute(in_map, AOI, origin_vtpk_path, out_part_vtpk):
+def execute(in_map, AOI, origin_vtpk, is_adv_vtpk, index_polygon, out_part_vtpk):
 
     arcpy.env.workspace = os.path.dirname(out_part_vtpk)
+    temp_workspace = None
 
-    # analyzing original_vtpk to get the scheme and index polygon feature class
-    aux_paras = analysis_original_vtpk(origin_vtpk_path)
-    index_polygon = aux_paras[0]
-    tile_scheme = aux_paras[1]
-    service_type = aux_paras[2]
-    temp_workspace = aux_paras[3]
-    arcpy.AddMessage("Analyzing origin vtpk succesfully! "
-                     +"\n - "+index_polygon+"\n - "+tile_scheme+"\n - "+service_type)
+    # Update 2018-9-20, add support for user defined xml file
+    if is_adv_vtpk:
+        # analyzing original_vtpk to get the scheme and index polygon feature class
+        adv_aux_paras = analysis_original_vtpk(origin_vtpk)
+        index_polygon = adv_aux_paras[0]
+        tile_scheme = adv_aux_paras[1]
+        service_type = adv_aux_paras[2]
+        temp_workspace = adv_aux_paras[3]
+        arcpy.AddMessage("- Analyzing origin vtpk succesfully! "
+                         + "\n - " + index_polygon + "\n - " + tile_scheme + "\n - " + service_type)
+    else:
+        common_aux_paras = get_tile_scheme_and_index(in_map)
+        tile_scheme = common_aux_paras[0]
+        service_type = common_aux_paras[1]
+        arcpy.AddMessage("- Servcie Type: " + service_type)
 
     # Calculating affected bundle indexes
     bundle_index_polygons = calculate_affected_indexes_layer(AOI, index_polygon)
@@ -180,7 +221,7 @@ def execute(in_map, AOI, origin_vtpk_path, out_part_vtpk):
     if os.path.exists(out_part_vtpk) and arcpy.Exists(out_part_vtpk):
         arcpy.Delete_management(os.path.join(arcpy.env.workspace, bundle_index_polygons))
         arcpy.AddMessage("Temp Data deleted - " + bundle_index_polygons)
-    if os.path.exists(out_part_vtpk):
+    if os.path.exists(out_part_vtpk) and temp_workspace is not None:
         shutil.rmtree(temp_workspace)
         arcpy.AddMessage("Temp Data deleted - " + temp_workspace)
